requirements.txt
tensorflow==1.9.0
numpy==1.14.5
six==1.11.0
opencv_contrib_python==3.4.1.15
Pillow==5.3.0
ImageTk
image
